컨테이너 컴포넌트 모두 제거
