+++ src/components/MakeTeam/MakeTeamPage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 import {useHistory, Route, useRouteMatch} from 'react-router-dom';
-import { MakeTeamContainers, MakeTeamDoneContainer} from '../index';
+import { MakeTeam, MakeTeamDone} from '../index';
 
 function MakeTeamPage(){
     const history = useHistory();
@@ -10,8 +10,8 @@ function MakeTeamPage(){
     
     return (
         <div css={style}>
-            <Route path={`${match.path}`} exact render={()=> <MakeTeamContainers/>}/>
-            <Route path={`${match.path}/done`} exact render={()=> <MakeTeamDoneContainer/>}/>
+            <Route path={`${match.path}`} exact render={()=> <MakeTeam/>}/>
+            <Route path={`${match.path}/done`} exact render={()=> <MakeTeamDone/>}/>
         </div>
     );
 }
