+++ src/components/index.ts
@@ -1,6 +1,3 @@
-import MakeTeamContainers from "./MakeTeam/MakeTeamContainer"
-import MakeTeamDoneContainer from "./MakeTeam/MakeTeamDoneContainer"
-import MemberInfoContainer from "./MyTeam/MemberInfoContainer"
 import GatheringFilter from "./Gathering/GatheringFilter"
 import GatheringList from "./Gathering/GatheringList"
 import GatheringCard from "./Gathering/GatheringCard"
@@ -59,9 +56,6 @@ import LoadingMark from "./LoadingMark"
 import LoadingSpinner from "./LoadingSpinner"
 
 export {
-    MakeTeamContainers,
-    MakeTeamDoneContainer,
-    MemberInfoContainer,
     GatheringFilter,
     GatheringList,
     GatheringCard,
